opol/stack/flows/x_dev/bundestagapidaily/requirements.txt
onnxruntime
prettyprint
opol
prefect
scikit-learn
numpy
google-generativeai
instructor
transformers
torch
matplotlib
einops
sentence-transformers
